Fixed toField

+++ sources/utils/Number.cpp
@@ -101,7 +101,7 @@ void Number::toField(unsigned int modulo) {
     }
     // Convert the value back to digits
     digits.clear();
-    int nDigits = floor(log10(value)) + 1;
+    int nDigits = int(floor(log10(value)) + 1);
     for (int i = 0; i < nDigits; ++i) {
         digits.push_back(value % 10);
         value /= 10;
